+++ src/modules/filters/utf8bidireorder.cpp
@@ -50,7 +50,7 @@ UTF8BiDiReorder::~UTF8BiDiReorder() {
 char UTF8BiDiReorder::processText(SWBuf &text, const SWKey *key, const SWModule *module)
 {
         UChar *ustr, *ustr2;
-	if ((unsigned long)key < 2)	// hack, we're en(1)/de(0)ciphering
+	if ((size_t)key < 2)	// hack, we're en(1)/de(0)ciphering
 		return -1;
         
         int32_t len = text.length();
